Added new CPS property framework.kube.launch.interval.milliseconds to solve #2045
Signed-off-by: Mike Cobbett 77053+techcobweb@users.noreply.github.com

Why ?
See issue Galasa Ecosystem test runs are causing k8s nodes to run out of memory with a high enough throttle value #2045

added a new CPS property
framework.kube.launch.interval.milliseconds
The number of milliseconds which the test pod scheduler should delay between successive launches of test pods.

The default is CPSFacade.KUBE_LAUNCH_INTERVAL_MILLISECOND_DEFAULT_VALUE

Why do this ?

If we don't do this, then all the tests get scheduled on the same node, and the
node may run out of memory.

We assume that's because the usage statistics on a pod are not synchronized totally at
real-time, but have a lag in which they catch up. Hopefully this delay is greater
than the lag and when we actually schedule the next pod it gets evenly distributed over
the nodes which are available.

This may or may not be necessary if the scheduling policies in the cluster are changed.

This CPS property is dynamic, in that you don't need to re-start the test launcher pod for it to take effect.
It gets read every time the value is needed, and is not cached.

Any failure in the CPS will be logged and ignored, resulting in the default value being returned.