IMS TM Manager (#2148) (#199)
Signed-off-by: Ian Deeley <ian_deeley@uk.ibm.com>

Signed-off-by: Mike Cobbett 77053+techcobweb@users.noreply.github.com

Initial drop of the new IMS TM Manager. This contains what should be the final product code (subject to review and/or defects), build updates, and the first batch of JUnits.

Code is heavily based on the CICS TM Manager  - I basically just copied everything and then modified it. In many cases, this was just doing change/replace on "CICS"->"IMS", "TS"->"TM", "Region"->"System" (matching case), but there was also a lot of deletions. And the system logon code is very different.

There are multiple commits in this PR, and it should be easier to review the code in the order they were committed.

Matching issue [here](https://github.com/galasa-dev/projectmanagement/issues/2148)