CLI syntax documentation to be generated into a zip for further use by the docs build process (#295)
Signed-off-by: Mike Cobbett <77053+techcobweb@users.noreply.github.com>

# Why ?
The CLI docs have always been checked-in to the CLI source repo and viewed through github UI.

We want to generate the docs such that they are markdown inside the documentation source, when the whole thing can be built into a web site. 

So combining the .md files prior to building the docs is what we want to do.

This PR builds the .md zip of the CLI syntax documentation and zips it up, placing it in the dev.galasa.cli.docs maven artifact.