3270 manager: set initial cursor position and send up to 256 bytes of outbound data for SSCP-LU session screens (#301)
## Why?
For https://github.com/galasa-dev/projectmanagement/issues/2232

SSCP-LU session displays are unformatted and use an "initial cursor position" to determine the start of an input area. The initial cursor position should be set to one screen position after the last character in the SSCP-LU inbound data stream, or the start of the next line if the last character was a newline.

The input area has an outbound limit of 256 bytes of data, data entry is restricted from the initial cursor position to the following 255 screen positions or to the last screen position.

This PR updates the 3270 manager to take these attributes of SSCP-LU session displays into account when entering data into such displays.

## Changes
- An initial cursor position is now stored after processing inbound messages
- When building the outbound message bytes, SSCP-LU session screens will build messages starting from the initial cursor position onwards as described above
- Added an insert cursor instruction when converting SSCP-LU data streams into standard 3270 text, which moves the cursor to one position after the last character/the start of the next line in the event of the last character being a newline
  - Note: The 3270 manager already had code that processed unformatted screens in standard 3270 sessions (i.e. not SSCP-LU sessions). This existing processing has been left as-is, since those screens are intended to be sent in their entirety

---------

Signed-off-by: Eamonn Mansour <47121388+eamansour@users.noreply.github.com>